refactor: move some method from profile core
